requirements.txt
diffusers==0.32.2
nltk==3.9.1
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
spacy==3.8.5
torch==2.6.0
tqdm==4.67.1
urllib3==2.3.0
